checkpoint 39 autentication Guest error nilclass

+++ app/controllers/posts_controller.rb
@@ -1,12 +1,13 @@
 class PostsController < ApplicationController
-  
   def index
-    @posts = Post.all
-    authorize @posts
+    #@posts = Post.all
+    @posts = policy_scope(Post)
+    #authorize @posts
   end
 
   def show
     @post = Post.find(params[:id])
+    authorize @post
   end
 
   def new
@@ -16,6 +17,7 @@ class PostsController < ApplicationController
   
   def create
     @post = current_user.posts.build(params.require(:post).permit(:title, :body))
+    authorize @post
     if @post.save
       flash[:notice] = "Post was saved."
       redirect_to @post
@@ -43,6 +45,8 @@ class PostsController < ApplicationController
   end
   
   def destroy
+    @post = Post.find(params[:id])
+    authorize @post
     @post = Post.find(params[:id])
     @post.destroy
     if @post.destroy
